Move the scoring functions out of lane.py and into external_search.py, since they deal directly with Elasticsearch.
