+++ tests/test_external_search.py
@@ -18,6 +18,10 @@ from . import (
 )
 
 from elasticsearch_dsl import Q
+from elasticsearch_dsl.function import (
+    ScriptScore,
+    RandomScore,
+)
 from elasticsearch_dsl.query import (
     Bool,
     Query as elasticsearch_dsl_query,
@@ -1621,6 +1625,67 @@ class TestFeaturedFacets(EndToEndSearchTest):
         self.best_seller_list.add_entry(self.featured_on_list, featured=True)
         self.best_seller_list.add_entry(self.not_featured_on_list)
 
+    def test_scoring_functions(self):
+        # Verify that FeaturedFacets sets appropriate scoring functions
+        # for ElasticSearch queries.
+        f = FeaturedFacets(minimum_featured_quality=0.55, random_seed=42)
+        filter = Filter()
+        f.modify_search_filter(filter)
+
+        # In most cases, there are three things that can boost a work's score.
+        [featurable, available_now, random] = f.scoring_functions(filter)
+
+        # It can be high-quality enough to be featured.
+        assert isinstance(featurable, ScriptScore)
+        source = filter.FEATURABLE_SCRIPT % dict(
+            cutoff=f.minimum_featured_quality ** 2, exponent=2
+        )
+        eq_(source, featurable.script['source'])
+
+        # It can be currently available.
+        availability_filter = available_now['filter']
+        eq_(
+            dict(nested=dict(
+                path='licensepools',
+                query=dict(term={'licensepools.available': True})
+            )),
+            availability_filter.to_dict()
+        )
+        eq_(5, available_now['weight'])
+
+        # It can get lucky.
+        assert isinstance(random, RandomScore)
+        eq_(42, random.seed)
+        eq_(1.1, random.weight)
+
+        # If the FeaturedFacets is set to be deterministic (which only happens
+        # in tests), the RandomScore is removed.
+        f.random_seed = filter.DETERMINISTIC
+        [featurable_2, available_now_2] = f.scoring_functions(filter)
+        eq_(featurable_2, featurable)
+        eq_(available_now_2, available_now)
+
+        # If custom lists are in play, it can also be featured on one
+        # of its custom lists.
+        filter.customlist_restriction_sets = [[1,2], [3]]
+        [featurable_2, available_now_2,
+         featured_on_list] = f.scoring_functions(filter)
+        eq_(featurable_2, featurable)
+        eq_(available_now_2, available_now)
+
+        # Any list will do -- the customlist restriction sets aren't
+        # relevant here.
+        featured_filter = featured_on_list['filter']
+        eq_(dict(
+            nested=dict(
+                path='customlists',
+                query=dict(bool=dict(
+                    must=[{'term': {'customlists.featured': True}},
+                          {'terms': {'customlists.list_id': [1, 2, 3]}}])))),
+            featured_filter.to_dict()
+        )
+        eq_(11, featured_on_list['weight'])
+
     def test_run(self):
 
         def assert_featured(description, worklist, facets, expect):
@@ -1633,7 +1698,7 @@ class TestFeaturedFacets(EndToEndSearchTest):
 
         worklist = WorkList()
         worklist.initialize(self._default_library)
-        facets = FeaturedFacets(1, random_seed=FeaturedFacets.DETERMINISTIC)
+        facets = FeaturedFacets(1, random_seed=Filter.DETERMINISTIC)
 
         # Even though hq_not_available is higher-quality than
         # featured_on_list, it shows up first because it's available
@@ -1668,7 +1733,7 @@ class TestFeaturedFacets(EndToEndSearchTest):
         # from being considered altogether. Basically all that matters
         # is availability.
         all_featured_facets = FeaturedFacets(
-            0, random_seed=FeaturedFacets.DETERMINISTIC
+            0, random_seed=Filter.DETERMINISTIC
         )
         assert_featured(
             "Works without considering quality",
@@ -1684,12 +1749,13 @@ class TestFeaturedFacets(EndToEndSearchTest):
         #
         # The random element is relatively small, so it mainly acts
         # to rearrange works whose scores were similar before.
-        random_facets = FeaturedFacets(1, random_seed=41)
+        random_facets = FeaturedFacets(1, random_seed=43)
         assert_featured(
             "Works permuted by a random seed",
             worklist, random_facets,
-            [self.hq_available_2, self.hq_available,
-             self.not_featured_on_list, self.hq_not_available,
+            [self.not_featured_on_list, self.hq_available_2,
+             self.hq_available,
+             self.hq_not_available,
              self.featured_on_list],
         )
